add null gene mutator
This gene mutator performs no action on the original hash. It is the
default gene mutator if none is provided to the GenePool initializer

You could also use a SimpleGeneMutator with a mutation rate of 0, I
suppose.

It could have been written as Proc.new({|x| x}), but I wanted to be more
explicit.