add mutate method to Gene
GA use combination and mutation to produce the next generation.
The mutate method takes a mutator as an argument. This can be a simple
Proc object that manipulates the gene information in some way to produce
a valid alternate. (i.e. it must conform to the gene template)
A simple Null mutator would be Proc.new({|x| x})