sort selection pool in descending order
This was sorted in ascending order - the lowest performing gene was
first, meaning it would be picked more often, ruining the selection
algorithm! Survival of the weakest!

The fix is to .reverse the sorted array. Saw an interesting article
about sorting by the -ve value to improve performance. If I'm worried
about performance, I'll look into it.