tweak gene selection algorithm
The min/max thing was a bug. I want to to be never more that 0.8, ofc.
Also tweak the probability to be a bit less top-heavy and allow a
slightly greater distribution of genes.

This gene selection algorithm should be extracted out into a separate
class at some point, and can be much more sophisticated. But this will
do for now.