improve optimisation tests
I haven't quite got my head around how to implement these test without
dependency-injecting ALL THE THINGS.
So, they don't pass all the time, because of the random nature of the
initial conditions and the non-deterministic nature of the algorithm.

This is not suitable for jenkins, ofc. But the GA now has a complete set
of working parts and will optimise to some degree. Thats the minimal set
of requirments met for this initial development branch.